pyproject.toml
[tool.poetry]
name = "catalog4-client"
version = "0.1.0"
description = ""
authors = ["sorochinskii <sorochinsky.alex@gmail.com>"]
license = "unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.2"
pandas = "^2.2.3"
reflex = "^0.6.5"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pydantic = "^2.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
